Fixed performance issue in LiveCharts.Wpf.DefaultLegend
Fixed a performance issue in DefaultLegend. PropertyChanged event for property "Series" was invoked every time, even if no item in the series list has changed. In charts with a huge amount of lines with a lot of data, this will trigger the redrawing of the legend so many times, that it will freeze the whole UI.